requirements.txt
django>=1.8.0
wheel==0.24.0
django-mptt
django-ckeditor